requirements.txt
streamlit
google-generativeai
pandas
numpy
python-dotenv
Pillow